Switched the VS project out for a stock Web App
