+++ BAMTriviaProject2/BAMTriviaProject2/Startup.cs
@@ -10,7 +10,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Project2
+namespace BAMTriviaProject2
 {
     public class Startup
     {
