+++ BAMTriviaProject2/BAMTriviaProject2/Views/Home/Index.cshtml
