+++ BAMTriviaProject2/BAMTriviaProject2/Views/_ViewImports.cshtml
@@ -1,0 +1,3 @@
+﻿@using BAMTriviaProject2
+@using BAMTriviaProject2.Models
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
