+++ BAMTriviaProject2/BAMTriviaProject2/appsettings.Development.json
