+++ BAMTriviaProject2/BAMTriviaProject2/appsettings.json
