--- BAMTriviaProject2/Project2/Views/_ViewImports.cshtml
@@ -1,3 +1,0 @@
-﻿@using Project2
-@using Project2.Models
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
